+++ src/System.Windows.Forms/src/System/Windows/Forms/Form.cs
@@ -31,7 +31,7 @@ namespace System.Windows.Forms
     DefaultEvent(nameof(Load)),
     InitializationEvent(nameof(Load)),
     ]
-    public class Form : ContainerControl
+    public partial class Form : ContainerControl
     {
 #if DEBUG
         static readonly BooleanSwitch AlwaysRestrictWindows = new BooleanSwitch("AlwaysRestrictWindows", "Always make Form classes behave as though they are restricted");
@@ -7292,62 +7292,6 @@ namespace System.Windows.Forms
                 default:
                     base.WndProc(ref m);
                     break;
-            }
-        }
-
-        /// <summary>
-        ///  Form control accessible object with UI Automation provider functionality.
-        ///  This inherits from the base ControlAccessibleObject
-        ///  to have all base functionality.
-        /// </summary>
-        internal class FormAccessibleObject : ControlAccessibleObject
-        {
-            private readonly Form _owner;
-
-            internal FormAccessibleObject(Form owner) : base(owner)
-            {
-                _owner = owner;
-            }
-
-            public override Rectangle Bounds
-            {
-                get
-                {
-                    // Get rectangle without a title bar
-                    Rectangle rectangle = _owner.GetToolNativeScreenRectangle();
-
-                    // Add a title bar height 
-                    int titleHeight = rectangle.Y - _owner.Top - 1;
-                    rectangle.Y -= titleHeight;
-                    rectangle.Height += titleHeight;
-
-                    return rectangle;
-                }
-            }
-
-            internal override bool IsIAccessibleExSupported()
-            {
-                if (_owner != null)
-                {
-                    return true;
-                }
-
-                return base.IsIAccessibleExSupported();
-            }
-
-            internal override bool IsPatternSupported(int patternId)
-            {
-                if (patternId == NativeMethods.UIA_LegacyIAccessiblePatternId)
-                {
-                    return true;
-                }
-
-                return base.IsPatternSupported(patternId);
-            }
-
-            internal override void SetValue(string newValue)
-            {
-                Value = newValue;
             }
         }
